Trait lasso::IntoReader

pub trait IntoReader<K = Spur>: Interner<K> where
    K: Key, {
    type Reader: Reader<K>;
    #[must_use]
    fn into_reader(self) -> Self::Reader
    where
        Self: 'static;
}

A generic interface over interners that can be turned into a Reader.

Associated Types

type Reader: Reader<K>

The type of Reader the interner will be converted into

Required methods

#[must_use]
fn into_reader(self) -> Self::Reader where
Self: 'static,

Consumes the current Interner and converts it into a Reader to allow contention-free access of the interner from multiple threads