Trait lasso::IntoReaderAndResolver

pub trait IntoReaderAndResolver<K = Spur>: IntoReader<K> + IntoResolver<K>where
    K: Key,{ }

A generic interface over interners that can be turned into both a Reader and a Resolver directly.