Trait lasso::Reader

pub trait Reader<K = Spur>: Resolver<K> {
    fn get(&self, val: &str) -> Option<K>;
    fn contains(&self, val: &str) -> bool;
}

A generic interface that allows using any underlying interner for both its reading and resolution capabilities, allowing both str -> key and key -> str lookups

Required Methods

fn get(&self, val: &str) -> Option<K>

Get a key for the given string value if it exists

fn contains(&self, val: &str) -> bool

Returns true if the current interner contains the given string