Trait lasso::Resolver

pub trait Resolver<K = Spur> {
    fn resolve<'a>(&'a self, key: &K) -> &'a str;

    fn try_resolve<'a>(&'a self, key: &K) -> Option<&'a str>;

    unsafe fn resolve_unchecked<'a>(&'a self, key: &K) -> &'a str;

    fn contains_key(&self, key: &K) -> bool;

    fn len(&self) -> usize;

    fn is_empty(&self) -> bool { ... }
}

A generic interface that allows using any underlying interner only for its resolution capabilities, allowing only key -> str lookups

Required methods

fn resolve<'a>(&'a self, key: &K) -> &'a str

Resolves the given key into a string

Panics

Panics if the key is not contained in the current Resolver

fn try_resolve<'a>(&'a self, key: &K) -> Option<&'a str>

Attempts to resolve the given key into a string, returning None if it cannot be found

unsafe fn resolve_unchecked<'a>(&'a self, key: &K) -> &'a str

Resolves a string by its key without preforming bounds checks

Safety

The key must be valid for the current Resolver

fn contains_key(&self, key: &K) -> bool

Returns true if the current interner contains the given key

fn len(&self) -> usize

Gets the number of currently interned strings

Provided methods

fn is_empty(&self) -> bool

Returns true if there are no currently interned strings